@@ -14,7 +14,7 @@ description: An app for getting around.
 version: 1.0.7+7
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
 
 #dependency_overrides:
 #  image: 2.0.7
